Move some diffusion utils to shared test_utils so they can be imported from tools
