+++ model/common/src/icon4py/model/common/dimension.py
@@ -60,6 +60,3 @@ KHalfOff = FieldOffset("KHalfOff", source=KHalfDim, target=(KHalfDim,))
 
 
 VerticalDim: TypeAlias = Union[KDim, KHalfDim]
-CellIndexDim = Dimension("CellIndex")
-EdgeIndexDim = Dimension("EdgeIndex")
-VertexIndexDim = Dimension("VertexIndex")
