+++ tools/src/icon4pytools/py2fgen/wrappers/diffusion.py
@@ -16,13 +16,10 @@ Fortran granule interfaces:
 - all arguments needed from external sources are passed.
 - passing of scalar types or fields of simple types
 """
-import cProfile
-import pstats
 
 from gt4py.next.common import Field
 from gt4py.next.ffront.fbuiltins import float64, int32
 from icon4py.model.atmosphere.diffusion.diffusion import (
-    Diffusion,
     DiffusionConfig,
     DiffusionParams,
     TurbulenceShearForcingType,
@@ -32,97 +29,18 @@ from icon4py.model.atmosphere.diffusion.diffusion_states import (
     DiffusionInterpolationState,
     DiffusionMetricState,
 )
-from icon4py.model.common import dimension as dims, settings
+from icon4py.model.common import dimension as dims
 from icon4py.model.common.constants import DEFAULT_PHYSICS_DYNAMICS_TIMESTEP_RATIO
 from icon4py.model.common.grid import geometry
-from icon4py.model.common.grid.icon import GlobalGridParams, IconGrid
+from icon4py.model.common.grid.icon import GlobalGridParams
 from icon4py.model.common.grid.vertical import VerticalGrid, VerticalGridConfig
 from icon4py.model.common.settings import device
 from icon4py.model.common.states.prognostic_state import PrognosticState
 from icon4py.model.common.test_utils.helpers import as_1D_sparse_field
-
 from icon4pytools.common.logger import setup_logger
-from icon4pytools.py2fgen.wrappers.common import construct_icon_grid
-
+from icon4pytools.py2fgen.wrappers import common
 
 logger = setup_logger(__name__)
-
-# global diffusion object
-diffusion_granule: Diffusion = Diffusion()
-
-# global profiler object
-profiler = cProfile.Profile()
-
-# global grid object
-icon_grid: IconGrid = None
-
-
-def profile_enable():
-    profiler.enable()
-
-
-def profile_disable():
-    profiler.disable()
-    stats = pstats.Stats(profiler)
-    stats.dump_stats(f"{__name__}.profile")
-
-
-def grid_init(
-    grid_id: str,
-    cell_starts: Field[[dims.CellIndexDim], int32],
-    cell_ends: Field[[dims.CellIndexDim], int32],
-    vertex_starts: Field[[dims.VertexIndexDim], int32],
-    vertex_ends: Field[[dims.VertexIndexDim], int32],
-    edge_starts: Field[[dims.EdgeIndexDim], int32],
-    edge_ends: Field[[dims.EdgeIndexDim], int32],
-    c2e: Field[[dims.CellDim, dims.C2EDim], int32],
-    e2c: Field[[dims.EdgeDim, dims.E2CDim], int32],
-    c2e2c: Field[[dims.CellDim, dims.C2E2CDim], int32],
-    e2c2e: Field[[dims.EdgeDim, dims.E2C2EDim], int32],
-    e2v: Field[[dims.EdgeDim, dims.E2VDim], int32],
-    v2e: Field[[dims.VertexDim, dims.V2EDim], int32],
-    v2c: Field[[dims.VertexDim, dims.V2CDim], int32],
-    e2c2v: Field[[dims.EdgeDim, dims.E2C2VDim], int32],
-    c2v: Field[[dims.CellDim, dims.C2VDim], int32],
-    c2e2c2e: Field[[dims.CellDim, dims.C2E2C2EDim], int32],
-    global_root: int32,
-    global_level: int32,
-    num_vertices: int32,
-    num_cells: int32,
-    num_edges: int32,
-    vertical_size: int32,
-    limited_area: bool,
-):
-    global icon_grid
-
-    global_grid_params = GlobalGridParams(level=global_level, root=global_root)
-
-    icon_grid = construct_icon_grid(
-        grid_id=grid_id,
-        global_grid_params=global_grid_params,
-        num_vertices=num_vertices,
-        num_cells=num_cells,
-        num_edges=num_edges,
-        vertical_size=vertical_size,
-        limited_area=limited_area,
-        on_gpu=True if settings.device == "GPU" else False,
-        cell_starts=cell_starts,
-        cell_ends=cell_ends,
-        vertex_starts=vertex_starts,
-        vertex_ends=vertex_ends,
-        edge_starts=edge_starts,
-        edge_ends=edge_ends,
-        c2e=c2e,
-        e2c=e2c,
-        c2e2c=c2e2c,
-        e2c2e=e2c2e,
-        e2v=e2v,
-        v2e=v2e,
-        v2c=v2c,
-        e2c2v=e2c2v,
-        c2v=c2v,
-        c2e2c2e=c2e2c2e,
-    )
 
 
 def diffusion_init(
@@ -244,7 +162,7 @@ def diffusion_init(
 
     # Vertical grid config
     vertical_config = VerticalGridConfig(
-        num_levels=icon_grid.num_levels,
+        num_levels=common.GLOBAL_STATE["icon_grid"].num_levels,
         lowest_layer_thickness=lowest_layer_thickness,
         model_top_height=model_top_height,
         stretch_factor=stretch_factor,
@@ -264,8 +182,10 @@ def diffusion_init(
         mask_hdiff=mask_hdiff,
         theta_ref_mc=theta_ref_mc,
         wgtfac_c=wgtfac_c,
-        zd_intcoef=zd_intcoef,  # todo: for icon integration? flatten_first_two_dims(dims.CECDim, dims.KDim, field=zd_intcoef),
-        zd_vertoffset=zd_vertoffset,  # todo: for icon integration? flatten_first_two_dims(dims.CECDim, dims.KDim, field=zd_vertoffset),
+        zd_intcoef=zd_intcoef,
+        # todo: for icon integration? flatten_first_two_dims(dims.CECDim, dims.KDim, field=zd_intcoef),
+        zd_vertoffset=zd_vertoffset,
+        # todo: for icon integration? flatten_first_two_dims(dims.CECDim, dims.KDim, field=zd_vertoffset),
         zd_diffcoef=zd_diffcoef,
     )
 
@@ -282,8 +202,8 @@ def diffusion_init(
     )
 
     # Initialize the diffusion granule
-    diffusion_granule.init(
-        grid=icon_grid,
+    common.GLOBAL_STATE["diffusion_granule"].init(
+        grid=common.GLOBAL_STATE["icon_grid"],
         config=config,
         params=diffusion_params,
         vertical_grid=vertical_params,
@@ -324,12 +244,12 @@ def diffusion_run(
     )
 
     if linit:
-        diffusion_granule.initial_run(
+        common.GLOBAL_STATE["diffusion_granule"].initial_run(
             diagnostic_state,
             prognostic_state,
             dtime,
         )
     else:
-        diffusion_granule.run(
+        common.GLOBAL_STATE["diffusion_granule"].run(
             prognostic_state=prognostic_state, diagnostic_state=diagnostic_state, dtime=dtime
         )
